chore: Refactor `cache` into `core.
