+++ packages/analytics/package.json
@@ -48,7 +48,6 @@
 		"src"
 	],
 	"dependencies": {
-		"@aws-amplify/cache": "5.1.3",
 		"@aws-amplify/core": "5.5.2",
 		"@aws-sdk/client-firehose": "3.6.1",
 		"@aws-sdk/client-kinesis": "3.6.1",
