+++ packages/api-graphql/package.json
@@ -52,7 +52,6 @@
   "dependencies": {
     "@aws-amplify/api-rest": "3.3.2",
     "@aws-amplify/auth": "5.5.3",
-    "@aws-amplify/cache": "5.1.3",
     "@aws-amplify/core": "5.5.2",
     "@aws-amplify/pubsub": "5.3.3",
     "graphql": "15.8.0",
