+++ packages/api-graphql/src/internals/InternalGraphQLAPI.ts
@@ -16,10 +16,10 @@ import {
 	CustomUserAgentDetails,
 	getAmplifyUserAgent,
 	INTERNAL_AWS_APPSYNC_REALTIME_PUBSUB_PROVIDER,
+	Cache,
 } from '@aws-amplify/core';
 import { InternalPubSub } from '@aws-amplify/pubsub/internals';
 import { Auth } from '@aws-amplify/auth';
-import { Cache } from '@aws-amplify/cache';
 import {
 	GraphQLAuthError,
 	GraphQLOptions,
