+++ packages/api/webpack.config.js
@@ -9,7 +9,6 @@ module.exports = {
 		'graphql/language/printer',
 		{
 			'@aws-amplify/auth': 'aws_amplify_auth',
-			'@aws-amplify/cache': 'aws_amplify_cache',
 			'@aws-amplify/core': 'aws_amplify_core'
 		}
 	],
