+++ packages/auth/webpack.config.js
@@ -3,7 +3,6 @@ module.exports = {
 		'aws-amplify-auth.min': './lib-esm/index.js',
 	},
 	externals: ['react-native', {
-		'@aws-amplify/cache': 'aws_amplify_cache',
 		'@aws-amplify/core': 'aws_amplify_core'
 	}],
 	output: {
