+++ packages/aws-amplify/package.json
@@ -43,7 +43,6 @@
     "@aws-amplify/analytics": "6.3.2",
     "@aws-amplify/api": "5.3.3",
     "@aws-amplify/auth": "5.5.3",
-    "@aws-amplify/cache": "5.1.3",
     "@aws-amplify/core": "5.5.2",
     "@aws-amplify/datastore": "4.6.3",
     "@aws-amplify/geo": "2.1.3",
