+++ packages/aws-amplify/src/index.ts
@@ -21,7 +21,7 @@ export {
 	syncExpression,
 } from '@aws-amplify/datastore';
 export { PubSub } from '@aws-amplify/pubsub';
-export { Cache } from '@aws-amplify/cache';
+export { Cache } from '@aws-amplify/core';
 export { Interactions } from '@aws-amplify/interactions';
 export { Notifications } from '@aws-amplify/notifications';
 export { Predictions } from '@aws-amplify/predictions';
