--- packages/cache/build.js
@@ -1,7 +1,0 @@
-// Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
-// SPDX-License-Identifier: Apache-2.0
-'use strict';
-
-const build = require('../../scripts/build');
-
-build(process.argv[2], process.argv[3]);
