--- packages/cache/tslint.json
@@ -1,45 +1,0 @@
-{
-	"defaultSeverity": "error",
-	"plugins": ["prettier"],
-	"extends": [],
-	"jsRules": {},
-	"rules": {
-		"prefer-const": true,
-		"max-line-length": [true, 120],
-		"no-empty-interface": true,
-		"no-var-keyword": true,
-		"object-literal-shorthand": true,
-		"no-eval": true,
-		"space-before-function-paren": [
-			true,
-			{
-				"anonymous": "never",
-				"named": "never"
-			}
-		],
-		"no-parameter-reassignment": true,
-		"align": [true, "parameters"],
-		"no-duplicate-imports": true,
-		"one-variable-per-declaration": [false, "ignore-for-loop"],
-		"triple-equals": [true, "allow-null-check"],
-		"comment-format": [true, "check-space"],
-		"indent": [false],
-		"whitespace": [
-			false,
-			"check-branch",
-			"check-decl",
-			"check-operator",
-			"check-preblock"
-		],
-		"eofline": true,
-		"variable-name": [
-			true,
-			"check-format",
-			"allow-pascal-case",
-			"allow-snake-case",
-			"allow-leading-underscore"
-		],
-		"semicolon": [true, "always", "ignore-interfaces"]
-	},
-	"rulesDirectory": []
-}
