--- packages/cache/webpack.config.dev.js
@@ -1,6 +1,0 @@
-var config = require('./webpack.config.js');
-
-var entry = {
-	'aws-amplify-cache': './lib-esm/index.js',
-};
-module.exports = Object.assign(config, { entry, mode: 'development' });
