--- packages/cache/webpack.config.js
@@ -1,41 +1,0 @@
-module.exports = {
-	entry: {
-		'aws-amplify-cache.min': './lib-esm/index.js',
-	},
-	externals: [{ '@aws-amplify/core': 'aws_amplify_core' }, 'react-native'],
-	output: {
-		filename: '[name].js',
-		path: __dirname + '/dist',
-		library: 'aws_amplify_cache',
-		libraryTarget: 'umd',
-		umdNamedDefine: true,
-		globalObject: 'this',
-		devtoolModuleFilenameTemplate: require('../aws-amplify/webpack-utils')
-			.devtoolModuleFilenameTemplate,
-	},
-	// Enable sourcemaps for debugging webpack's output.
-	devtool: 'source-map',
-	resolve: {
-		extensions: ['.js', '.json'],
-	},
-	mode: 'production',
-	module: {
-		rules: [
-			// All output '.js' files will have any sourcemaps re-processed by 'source-map-loader'.
-			//{ enforce: 'pre', test: /\.js$/, loader: 'source-map-loader' },
-			{
-				test: /\.js?$/,
-				exclude: /node_modules/,
-				use: [
-					'babel-loader',
-					{
-						loader: 'babel-loader',
-						options: {
-							presets: ['@babel/preset-env'],
-						},
-					},
-				],
-			},
-		],
-	},
-};
