+++ packages/core/__tests__/Cache/BrowserStorageCache-unit-test.ts
@@ -1,9 +1,20 @@
+import { CacheConfig, CacheItem } from '../../src/Cache/types/Cache';
+import { defaultConfig, getByteLength } from '../../src/Cache/Utils/CacheUtils';
 import {
-	InMemoryCache as cache,
-	InMemoryCacheClass,
-} from '../src/InMemoryCache';
-import { defaultConfig, getByteLength } from '../src/Utils/CacheUtils';
-import { CacheConfig, CacheItem, CacheItemOptions } from '../src/types/Cache';
+	BrowserStorageCache as cache,
+	BrowserStorageCacheClass,
+} from '../../src/Cache/BrowserStorageCache';
+
+const config: CacheConfig = {
+	capacityInBytes: 3000,
+	itemMaxSize: 800,
+	defaultTTL: 3000000,
+	defaultPriority: 5,
+	warningThreshold: 0.8,
+	storage: window.localStorage,
+};
+
+cache.configure(config);
 
 function getItemSize(value: string): number {
 	const currTime: Date = new Date();
@@ -22,30 +33,24 @@ function getItemSize(value: string): number {
 	return ret.byteSize;
 }
 
-const config: CacheConfig = {
-	capacityInBytes: 3000,
-	itemMaxSize: 600,
-	defaultTTL: 3000000,
-	defaultPriority: 5,
-	warningThreshold: 0.8,
-	storage: window.localStorage,
-};
-
-cache.configure(config);
-
-describe('InMemoryCache', () => {
-	const cache_size = config['capacityInBytes'] || defaultConfig.capacityInBytes;
-	const default_ttl = config['defaultTTL'] || defaultConfig.defaultTTL;
-	const item_max_size = config['itemMaxSize'] || defaultConfig.itemMaxSize;
+describe('BrowserStorageCache', () => {
+	const cache_size = config.capacityInBytes || defaultConfig.capacityInBytes;
+	const default_ttl = config.defaultTTL || defaultConfig.defaultTTL;
+	const item_max_size = config.itemMaxSize || defaultConfig.itemMaxSize;
+	const warningThreshold =
+		config.warningThreshold || defaultConfig.warningThreshold;
 
 	let regularItem: string = '';
-	// item size would be 339 Byte
 	for (let i = 0; i < item_max_size / 2; i++) {
 		regularItem += 'a';
 	}
 
 	const regularItemSize: number = getItemSize(regularItem);
+
 	const maxItemNum: number = Math.floor(cache_size / regularItemSize);
+	const itemsNeedToPop: number = Math.ceil(
+		(cache_size * (1 - warningThreshold)) / regularItemSize
+	);
 
 	if (maxItemNum > default_ttl) {
 		console.error('incorrect paratmeter for test!');
@@ -55,36 +60,44 @@ describe('InMemoryCache', () => {
 
 	describe('setItem test', () => {
 		test('put string, happy case', () => {
-			let key: string = 'a';
+			const key: string = 'a';
 			cache.setItem(key, regularItem);
-			expect(cache.getItem(key)).toBe(regularItem);
+			const ret = cache.getItem(key);
 
+			expect(ret).toBe(regularItem);
 			cache.clear();
 		});
 
 		test('put object, happy case', () => {
-			let key: string = 'a';
+			const key: string = 'a';
 			const item: object = { abc: 123, edf: 456 };
+
 			cache.setItem(key, item);
-			expect(cache.getItem(key)).toEqual(item);
+			const ret = cache.getItem(key);
+
+			expect(ret).toEqual(item);
 
 			cache.clear();
 		});
 
 		test('put number, happy case', () => {
-			let key: string = 'a';
-			const item: number = 12345;
+			const key: string = 'a';
+			const item: number = 1234;
 			cache.setItem(key, item);
-			expect(cache.getItem(key)).toBe(item);
+			const ret = cache.getItem(key);
+
+			expect(ret).toBe(item);
 
 			cache.clear();
 		});
 
 		test('put boolean, happy case', () => {
-			let key: string = 'a';
-			const item: boolean = false;
+			const key: string = 'a';
+			const item: boolean = true;
 			cache.setItem(key, item);
-			expect(cache.getItem(key)).toBe(item);
+			const ret = cache.getItem(key);
+
+			expect(ret).toBe(item);
 
 			cache.clear();
 		});
@@ -137,11 +150,14 @@ describe('InMemoryCache', () => {
 			const val2: string = 'cbaabc';
 
 			cache.setItem(key, val1);
-			const cacheSizeBefore = cache.getCacheCurSize();
-			expect(cache.getItem(key)).toBe(val1);
+			const cacheSizeBefore: number = cache.getCacheCurSize();
+			const ret1 = cache.getItem(key);
+			expect(ret1).toBe(val1);
+
 			cache.setItem(key, val2);
-			const cacheSizeAfter = cache.getCacheCurSize();
-			expect(cache.getItem(key)).toBe(val2);
+			const cacheSizeAfter: number = cache.getCacheCurSize();
+			const ret2 = cache.getItem(key);
+			expect(ret2).toBe(val2);
 
 			expect(cacheSizeAfter - cacheSizeBefore).toBe(
 				getItemSize(val2) - getItemSize(val1)
@@ -159,10 +175,10 @@ describe('InMemoryCache', () => {
 			// fill the cache
 			for (let i = 0; i < maxItemNum; i++) {
 				key = i.toString();
-				if (i == 0) {
-					cache.setItem(key, regularItem, { priority: 4 });
+				if (i < itemsNeedToPop) {
+					cache.setItem(key, regularItem);
 					keysPoped.push(key);
-				} else cache.setItem(key, regularItem, { priority: 3 });
+				} else cache.setItem(key, regularItem, { priority: priority });
 			}
 
 			for (let i = 0; i < keysPoped.length; i++) {
@@ -193,7 +209,7 @@ describe('InMemoryCache', () => {
 				dateSpy.mockImplementation(() => {
 					return 1434319925275 + i;
 				});
-				if (i == 0) {
+				if (i < itemsNeedToPop) {
 					keysPoped.push(key);
 				}
 				cache.setItem(key, regularItem);
@@ -202,10 +218,47 @@ describe('InMemoryCache', () => {
 			key = maxItemNum.toString();
 			cache.setItem(key, regularItem);
 
-			for (let i = 0; i < keysPoped.length; i++) {
-				expect(cache.getItem(keysPoped[i])).toBeNull();
+			for (let i = 0; i <= maxItemNum; i++) {
+				if (i < keysPoped.length) {
+					expect(cache.getItem(keysPoped[i])).toBeNull();
+				} else {
+					expect(cache.getItem(i.toString())).not.toBeNull();
+				}
 			}
 
+			cache.clear();
+			dateSpy.mockRestore();
+		});
+
+		test('wipe out expired items when cache is full and after that cache has enough room for the item', () => {
+			let key: string = 'a';
+			const dateSpy = jest.spyOn(Date.prototype, 'getTime');
+
+			for (let i = 0; i < maxItemNum; i++) {
+				key = i.toString();
+				dateSpy.mockImplementation(() => {
+					return 1434319925275 + i;
+				});
+				// set ttl to maxItemNum/2
+				cache.setItem(key, regularItem, {
+					expires: 1434319925275 + i + maxItemNum / 2,
+				});
+			}
+
+			dateSpy.mockImplementation(() => {
+				return 1434319925275 + maxItemNum;
+			});
+
+			key = maxItemNum.toString();
+			cache.setItem(key, regularItem);
+
+			for (let i = 0; i < maxItemNum; i++) {
+				if (i < (maxItemNum % 2 ? maxItemNum / 2 : maxItemNum / 2 + 1)) {
+					expect(cache.getItem(i.toString())).toBeNull();
+				} else {
+					expect(cache.getItem(i.toString())).not.toBeNull();
+				}
+			}
 			cache.clear();
 			dateSpy.mockRestore();
 		});
@@ -216,11 +269,12 @@ describe('InMemoryCache', () => {
 			let key: string = 'a';
 
 			cache.setItem(key, regularItem);
-			expect(cache.getItem(key)).toBe(regularItem);
+			const ret = cache.getItem(key);
+
 			cache.clear();
 		});
 
-		test('item get cleaned if expired when trying to get it', () => {
+		test('item get cleaned if it expires when trying to get it', () => {
 			const timeSpy = jest.spyOn(Date.prototype, 'getTime');
 			let key: string = 'a';
 
@@ -233,6 +287,7 @@ describe('InMemoryCache', () => {
 			}
 
 			expect(cache.getItem('0')).toBeNull();
+
 			timeSpy.mockRestore();
 			cache.clear();
 		});
@@ -243,7 +298,7 @@ describe('InMemoryCache', () => {
 
 		test('item get refreshed when fetched from cache', () => {
 			let key: string = 'a';
-
+			let keysPoped: string[] = [];
 			const timeSpy = jest.spyOn(Date.prototype, 'getTime');
 			for (let i = 0; i < maxItemNum; i++) {
 				key = i.toString();
@@ -251,18 +306,25 @@ describe('InMemoryCache', () => {
 					// to ensure no item is expired
 					return 1434319925275 + i * (default_ttl / (maxItemNum * 2));
 				});
+				if (i < itemsNeedToPop) {
+					keysPoped.push(key);
+				}
 				cache.setItem(key, regularItem);
 			}
 
 			// refreshed
-			cache.getItem('0');
-
+			cache.getItem(keysPoped[0]);
 			key = maxItemNum.toString();
 			cache.setItem(key, regularItem);
 
-			expect(cache.getItem('0')).not.toBeNull();
-			expect(cache.getItem('1')).toBeNull();
-			//myHandler.showTheList();
+			for (let i = 0; i < keysPoped.length; i++) {
+				if (i == 0) {
+					expect(cache.getItem(keysPoped[0])).not.toBeNull();
+				} else {
+					expect(cache.getItem(keysPoped[i])).toBeNull();
+				}
+			}
+
 			cache.clear();
 			timeSpy.mockRestore();
 		});
@@ -283,7 +345,6 @@ describe('InMemoryCache', () => {
 			expect(cache.getItem('a')).toBe(25);
 
 			cache.clear();
-			//expect(cache.getItem('a', callback)).toBe(5);
 		});
 
 		test('output a console warning and return null if invalid keys', () => {
@@ -311,7 +372,7 @@ describe('InMemoryCache', () => {
 	});
 
 	describe('clear test', () => {
-		test('clear the cache', () => {
+		test('clear the cache, including the CurSize key', () => {
 			let key: string = 'a';
 			for (let i = 0; i < maxItemNum / 2; i++) {
 				key = i.toString();
@@ -323,6 +384,7 @@ describe('InMemoryCache', () => {
 				key = i.toString();
 				expect(cache.getItem(key)).not.toBeNull();
 			}
+			expect(cache.getCacheCurSize()).not.toBe(0);
 
 			cache.clear();
 
@@ -331,44 +393,51 @@ describe('InMemoryCache', () => {
 				key = i.toString();
 				expect(cache.getItem(key)).toBeNull();
 			}
+			expect(cache.getCacheCurSize()).toBe(0);
+		});
+
+		test("will not remove other users' item", () => {
+			window.sessionStorage.setItem('others-testb', 'abc');
+
+			cache.setItem('a', 'abc');
+			cache.clear();
+
+			expect(cache.getItem('a')).toBeNull();
+			expect(window.sessionStorage.getItem('others-testb')).not.toBeNull();
 		});
 	});
 
-	describe('getItemCursize test', () => {
+	describe('getCacheCurSize test', () => {
 		test('return 0 if cache is empty', () => {
 			expect(cache.getCacheCurSize()).toBe(0);
 		});
 
 		test('return cache currrent size if not empty', () => {
-			let key: string = 'a';
 			for (let i = 0; i < maxItemNum; i++) {
-				key = i.toString();
+				let key = i.toString();
 				cache.setItem(key, regularItem);
 			}
 
 			expect(cache.getCacheCurSize()).toBe(regularItemSize * maxItemNum);
-
 			cache.clear();
-			console.log(String.fromCharCode(0xdc33));
 		});
 	});
 
 	describe('getAllKeys test', () => {
-		test('happy case', async () => {
-			await cache.setItem('a', 123);
-			await cache.setItem('b', 'abc');
-			await cache.setItem('c', { abc: 123 });
+		test('happy case', () => {
+			cache.setItem('a', 123);
+			cache.setItem('b', 'abc');
+			cache.setItem('c', { abc: 123 });
 
-			expect.assertions(1);
-			expect(await cache.getAllKeys()).toEqual(['a', 'b', 'c']);
-			await cache.clear();
+			expect(cache.getAllKeys()).toEqual(['a', 'b', 'c']);
+			cache.clear();
 		});
 	});
 
 	describe('createInstance', () => {
 		test('happy case, return new instance', () => {
 			expect(cache.createInstance({ keyPrefix: 'abc' })).toBeInstanceOf(
-				InMemoryCacheClass
+				BrowserStorageCacheClass
 			);
 		});
 	});
