+++ packages/core/__tests__/Cache/InMemoryStorageCache-unit-test.ts
@@ -1,20 +1,9 @@
-import { CacheConfig, CacheItem } from '../src/types/Cache';
-import { defaultConfig, getByteLength } from '../src/Utils/CacheUtils';
 import {
-	BrowserStorageCache as cache,
-	BrowserStorageCacheClass,
-} from '../src/BrowserStorageCache';
-
-const config: CacheConfig = {
-	capacityInBytes: 3000,
-	itemMaxSize: 800,
-	defaultTTL: 3000000,
-	defaultPriority: 5,
-	warningThreshold: 0.8,
-	storage: window.localStorage,
-};
-
-cache.configure(config);
+	InMemoryCache as cache,
+	InMemoryCacheClass,
+} from '../../src/Cache/InMemoryCache';
+import { defaultConfig, getByteLength } from '../../src/Cache/Utils/CacheUtils';
+import { CacheConfig, CacheItem } from '../../src/Cache/types/Cache';
 
 function getItemSize(value: string): number {
 	const currTime: Date = new Date();
@@ -33,24 +22,30 @@ function getItemSize(value: string): number {
 	return ret.byteSize;
 }
 
-describe('BrowserStorageCache', () => {
-	const cache_size = config.capacityInBytes || defaultConfig.capacityInBytes;
-	const default_ttl = config.defaultTTL || defaultConfig.defaultTTL;
-	const item_max_size = config.itemMaxSize || defaultConfig.itemMaxSize;
-	const warningThreshold =
-		config.warningThreshold || defaultConfig.warningThreshold;
+const config: CacheConfig = {
+	capacityInBytes: 3000,
+	itemMaxSize: 600,
+	defaultTTL: 3000000,
+	defaultPriority: 5,
+	warningThreshold: 0.8,
+	storage: window.localStorage,
+};
+
+cache.configure(config);
+
+describe('InMemoryCache', () => {
+	const cache_size = config['capacityInBytes'] || defaultConfig.capacityInBytes;
+	const default_ttl = config['defaultTTL'] || defaultConfig.defaultTTL;
+	const item_max_size = config['itemMaxSize'] || defaultConfig.itemMaxSize;
 
 	let regularItem: string = '';
+	// item size would be 339 Byte
 	for (let i = 0; i < item_max_size / 2; i++) {
 		regularItem += 'a';
 	}
 
 	const regularItemSize: number = getItemSize(regularItem);
-
 	const maxItemNum: number = Math.floor(cache_size / regularItemSize);
-	const itemsNeedToPop: number = Math.ceil(
-		(cache_size * (1 - warningThreshold)) / regularItemSize
-	);
 
 	if (maxItemNum > default_ttl) {
 		console.error('incorrect paratmeter for test!');
@@ -60,44 +55,36 @@ describe('BrowserStorageCache', () => {
 
 	describe('setItem test', () => {
 		test('put string, happy case', () => {
-			const key: string = 'a';
+			let key: string = 'a';
 			cache.setItem(key, regularItem);
-			const ret = cache.getItem(key);
+			expect(cache.getItem(key)).toBe(regularItem);
 
-			expect(ret).toBe(regularItem);
 			cache.clear();
 		});
 
 		test('put object, happy case', () => {
-			const key: string = 'a';
+			let key: string = 'a';
 			const item: object = { abc: 123, edf: 456 };
-
 			cache.setItem(key, item);
-			const ret = cache.getItem(key);
-
-			expect(ret).toEqual(item);
+			expect(cache.getItem(key)).toEqual(item);
 
 			cache.clear();
 		});
 
 		test('put number, happy case', () => {
-			const key: string = 'a';
-			const item: number = 1234;
+			let key: string = 'a';
+			const item: number = 12345;
 			cache.setItem(key, item);
-			const ret = cache.getItem(key);
-
-			expect(ret).toBe(item);
+			expect(cache.getItem(key)).toBe(item);
 
 			cache.clear();
 		});
 
 		test('put boolean, happy case', () => {
-			const key: string = 'a';
-			const item: boolean = true;
+			let key: string = 'a';
+			const item: boolean = false;
 			cache.setItem(key, item);
-			const ret = cache.getItem(key);
-
-			expect(ret).toBe(item);
+			expect(cache.getItem(key)).toBe(item);
 
 			cache.clear();
 		});
@@ -150,14 +137,11 @@ describe('BrowserStorageCache', () => {
 			const val2: string = 'cbaabc';
 
 			cache.setItem(key, val1);
-			const cacheSizeBefore: number = cache.getCacheCurSize();
-			const ret1 = cache.getItem(key);
-			expect(ret1).toBe(val1);
-
+			const cacheSizeBefore = cache.getCacheCurSize();
+			expect(cache.getItem(key)).toBe(val1);
 			cache.setItem(key, val2);
-			const cacheSizeAfter: number = cache.getCacheCurSize();
-			const ret2 = cache.getItem(key);
-			expect(ret2).toBe(val2);
+			const cacheSizeAfter = cache.getCacheCurSize();
+			expect(cache.getItem(key)).toBe(val2);
 
 			expect(cacheSizeAfter - cacheSizeBefore).toBe(
 				getItemSize(val2) - getItemSize(val1)
@@ -175,10 +159,10 @@ describe('BrowserStorageCache', () => {
 			// fill the cache
 			for (let i = 0; i < maxItemNum; i++) {
 				key = i.toString();
-				if (i < itemsNeedToPop) {
-					cache.setItem(key, regularItem);
+				if (i == 0) {
+					cache.setItem(key, regularItem, { priority: 4 });
 					keysPoped.push(key);
-				} else cache.setItem(key, regularItem, { priority: priority });
+				} else cache.setItem(key, regularItem, { priority: 3 });
 			}
 
 			for (let i = 0; i < keysPoped.length; i++) {
@@ -209,7 +193,7 @@ describe('BrowserStorageCache', () => {
 				dateSpy.mockImplementation(() => {
 					return 1434319925275 + i;
 				});
-				if (i < itemsNeedToPop) {
+				if (i == 0) {
 					keysPoped.push(key);
 				}
 				cache.setItem(key, regularItem);
@@ -218,47 +202,10 @@ describe('BrowserStorageCache', () => {
 			key = maxItemNum.toString();
 			cache.setItem(key, regularItem);
 
-			for (let i = 0; i <= maxItemNum; i++) {
-				if (i < keysPoped.length) {
-					expect(cache.getItem(keysPoped[i])).toBeNull();
-				} else {
-					expect(cache.getItem(i.toString())).not.toBeNull();
-				}
+			for (let i = 0; i < keysPoped.length; i++) {
+				expect(cache.getItem(keysPoped[i])).toBeNull();
 			}
 
-			cache.clear();
-			dateSpy.mockRestore();
-		});
-
-		test('wipe out expired items when cache is full and after that cache has enough room for the item', () => {
-			let key: string = 'a';
-			const dateSpy = jest.spyOn(Date.prototype, 'getTime');
-
-			for (let i = 0; i < maxItemNum; i++) {
-				key = i.toString();
-				dateSpy.mockImplementation(() => {
-					return 1434319925275 + i;
-				});
-				// set ttl to maxItemNum/2
-				cache.setItem(key, regularItem, {
-					expires: 1434319925275 + i + maxItemNum / 2,
-				});
-			}
-
-			dateSpy.mockImplementation(() => {
-				return 1434319925275 + maxItemNum;
-			});
-
-			key = maxItemNum.toString();
-			cache.setItem(key, regularItem);
-
-			for (let i = 0; i < maxItemNum; i++) {
-				if (i < (maxItemNum % 2 ? maxItemNum / 2 : maxItemNum / 2 + 1)) {
-					expect(cache.getItem(i.toString())).toBeNull();
-				} else {
-					expect(cache.getItem(i.toString())).not.toBeNull();
-				}
-			}
 			cache.clear();
 			dateSpy.mockRestore();
 		});
@@ -269,12 +216,11 @@ describe('BrowserStorageCache', () => {
 			let key: string = 'a';
 
 			cache.setItem(key, regularItem);
-			const ret = cache.getItem(key);
-
+			expect(cache.getItem(key)).toBe(regularItem);
 			cache.clear();
 		});
 
-		test('item get cleaned if it expires when trying to get it', () => {
+		test('item get cleaned if expired when trying to get it', () => {
 			const timeSpy = jest.spyOn(Date.prototype, 'getTime');
 			let key: string = 'a';
 
@@ -287,7 +233,6 @@ describe('BrowserStorageCache', () => {
 			}
 
 			expect(cache.getItem('0')).toBeNull();
-
 			timeSpy.mockRestore();
 			cache.clear();
 		});
@@ -298,7 +243,7 @@ describe('BrowserStorageCache', () => {
 
 		test('item get refreshed when fetched from cache', () => {
 			let key: string = 'a';
-			let keysPoped: string[] = [];
+
 			const timeSpy = jest.spyOn(Date.prototype, 'getTime');
 			for (let i = 0; i < maxItemNum; i++) {
 				key = i.toString();
@@ -306,25 +251,18 @@ describe('BrowserStorageCache', () => {
 					// to ensure no item is expired
 					return 1434319925275 + i * (default_ttl / (maxItemNum * 2));
 				});
-				if (i < itemsNeedToPop) {
-					keysPoped.push(key);
-				}
 				cache.setItem(key, regularItem);
 			}
 
 			// refreshed
-			cache.getItem(keysPoped[0]);
+			cache.getItem('0');
+
 			key = maxItemNum.toString();
 			cache.setItem(key, regularItem);
 
-			for (let i = 0; i < keysPoped.length; i++) {
-				if (i == 0) {
-					expect(cache.getItem(keysPoped[0])).not.toBeNull();
-				} else {
-					expect(cache.getItem(keysPoped[i])).toBeNull();
-				}
-			}
-
+			expect(cache.getItem('0')).not.toBeNull();
+			expect(cache.getItem('1')).toBeNull();
+			//myHandler.showTheList();
 			cache.clear();
 			timeSpy.mockRestore();
 		});
@@ -345,6 +283,7 @@ describe('BrowserStorageCache', () => {
 			expect(cache.getItem('a')).toBe(25);
 
 			cache.clear();
+			//expect(cache.getItem('a', callback)).toBe(5);
 		});
 
 		test('output a console warning and return null if invalid keys', () => {
@@ -372,7 +311,7 @@ describe('BrowserStorageCache', () => {
 	});
 
 	describe('clear test', () => {
-		test('clear the cache, including the CurSize key', () => {
+		test('clear the cache', () => {
 			let key: string = 'a';
 			for (let i = 0; i < maxItemNum / 2; i++) {
 				key = i.toString();
@@ -384,7 +323,6 @@ describe('BrowserStorageCache', () => {
 				key = i.toString();
 				expect(cache.getItem(key)).not.toBeNull();
 			}
-			expect(cache.getCacheCurSize()).not.toBe(0);
 
 			cache.clear();
 
@@ -393,51 +331,44 @@ describe('BrowserStorageCache', () => {
 				key = i.toString();
 				expect(cache.getItem(key)).toBeNull();
 			}
-			expect(cache.getCacheCurSize()).toBe(0);
-		});
-
-		test("will not remove other users' item", () => {
-			window.sessionStorage.setItem('others-testb', 'abc');
-
-			cache.setItem('a', 'abc');
-			cache.clear();
-
-			expect(cache.getItem('a')).toBeNull();
-			expect(window.sessionStorage.getItem('others-testb')).not.toBeNull();
 		});
 	});
 
-	describe('getCacheCurSize test', () => {
+	describe('getItemCursize test', () => {
 		test('return 0 if cache is empty', () => {
 			expect(cache.getCacheCurSize()).toBe(0);
 		});
 
 		test('return cache currrent size if not empty', () => {
+			let key: string = 'a';
 			for (let i = 0; i < maxItemNum; i++) {
-				let key = i.toString();
+				key = i.toString();
 				cache.setItem(key, regularItem);
 			}
 
 			expect(cache.getCacheCurSize()).toBe(regularItemSize * maxItemNum);
+
 			cache.clear();
+			console.log(String.fromCharCode(0xdc33));
 		});
 	});
 
 	describe('getAllKeys test', () => {
-		test('happy case', () => {
-			cache.setItem('a', 123);
-			cache.setItem('b', 'abc');
-			cache.setItem('c', { abc: 123 });
+		test('happy case', async () => {
+			await cache.setItem('a', 123);
+			await cache.setItem('b', 'abc');
+			await cache.setItem('c', { abc: 123 });
 
-			expect(cache.getAllKeys()).toEqual(['a', 'b', 'c']);
-			cache.clear();
+			expect.assertions(1);
+			expect(await cache.getAllKeys()).toEqual(['a', 'b', 'c']);
+			await cache.clear();
 		});
 	});
 
 	describe('createInstance', () => {
 		test('happy case, return new instance', () => {
 			expect(cache.createInstance({ keyPrefix: 'abc' })).toBeInstanceOf(
-				BrowserStorageCacheClass
+				InMemoryCacheClass
 			);
 		});
 	});
