+++ packages/core/__tests__/Cache/Utils/cacheList-unit-test.ts
@@ -1,4 +1,4 @@
-import CacheList from '../../src/Utils/CacheList';
+import CacheList from '../../../src/Cache/Utils/CacheList';
 
 describe('cacheList', () => {
 	describe('isEmpty', () => {
