+++ packages/core/__tests__/Cache/Utils/cacheUtils-unit-test.ts
@@ -1,4 +1,4 @@
-import { getByteLength, getCurrTime } from '../../src/Utils/CacheUtils';
+import { getByteLength } from '../../../src/Cache/Utils/CacheUtils';
 
 describe('CacheUtils', () => {
 	describe('getByteLength test', () => {
