+++ packages/core/package.json
@@ -12,7 +12,11 @@
 		"./lib/I18n/index.js",
 		"./lib/Credentials.js",
 		"./lib-esm/I18n/index.js",
-		"./lib-esm/Credentials.js"
+		"./lib-esm/Credentials.js",
+		"./lib/Cache/BrowserStorageCache.js",
+    "./lib/Cache/AsyncStorageCache.js",
+    "./lib-esm/Cache/BrowserStorageCache.js",
+    "./lib-esm/Cache/AsyncStorageCache.js"
 	],
 	"scripts": {
 		"test": "npm run lint && jest -w 1 --coverage",
@@ -36,7 +40,8 @@
 		"./lib/index": "./lib-esm/index.js",
 		"./lib-esm/ClientDevice": "./lib-esm/ClientDevice/reactnative.js",
 		"./lib-esm/RNComponents": "./lib-esm/RNComponents/reactnative.js",
-		"./lib-esm/StorageHelper": "./lib-esm/StorageHelper/reactnative.js"
+		"./lib-esm/StorageHelper": "./lib-esm/StorageHelper/reactnative.js",
+		"./lib-esm/Cache": "./lib-esm/Cache/reactnative.js"
 	},
 	"repository": {
 		"type": "git",
@@ -147,7 +152,19 @@
 			"path": "./lib-esm/clients/middleware/signing/signer/signatureV4/index.js",
 			"import": "{ presignUrl }",
 			"limit": "6.3 kB"
-		}
+		},
+		{
+      "name": "Cache (default browser storage)",
+      "path": "./lib-esm/index.js",
+      "import": "{ Cache }",
+      "limit": "4 kB"
+    },
+    {
+      "name": "Cache (in-memory)",
+      "path": "./lib-esm/index.js",
+      "import": "{ InMemoryCache }",
+      "limit": "3.7 kB"
+    }
 	],
 	"jest": {
 		"globals": {
@@ -170,6 +187,10 @@
 			"json",
 			"jsx"
 		],
+		"setupFiles": [
+      "./__mocks__/SessionStorage.js",
+      "./__mocks__/LocalStorage.js"
+    ],
 		"testEnvironment": "jsdom",
 		"coverageThreshold": {
 			"global": {
