+++ packages/core/src/index.ts
@@ -98,6 +98,15 @@ export {
 	urlSafeEncode,
 } from './Util';
 
+// Cache exports
+import { BrowserStorageCache } from './Cache/BrowserStorageCache';
+export { InMemoryCache } from './Cache/InMemoryCache';
+export { CacheConfig } from './Cache/types';
+export { BrowserStorageCache };
+
+// Standard `Cache` export to maintain interoperability with React Native
+export { BrowserStorageCache as Cache };
+
 /**
  * @deprecated use named import
  */
