+++ packages/notifications/package.json
@@ -49,7 +49,6 @@
 		"src"
 	],
 	"dependencies": {
-		"@aws-amplify/cache": "5.1.3",
 		"@aws-amplify/core": "5.5.2",
 		"@aws-amplify/rtn-push-notification": "1.1.1",
 		"lodash": "^4.17.21",
