+++ packages/notifications/webpack.config.js
@@ -5,7 +5,6 @@ module.exports = {
 	externals: [
 		'react-native',
 		{
-			'@aws-amplify/cache': 'aws_amplify_cache',
 			'@aws-amplify/core': 'aws_amplify_core',
 		},
 		'@aws-sdk/client-pinpoint',
