+++ packages/pubsub/__tests__/AWSAppSyncRealTimeProvider.test.ts
@@ -10,9 +10,8 @@ jest.mock('@aws-amplify/core', () => ({
 }));
 
 import Observable from 'zen-observable-ts';
-import { Reachability, Credentials, Logger, Signer } from '@aws-amplify/core';
+import { Reachability, Credentials, Logger, Signer, Cache } from '@aws-amplify/core';
 import { Auth } from '@aws-amplify/auth';
-import { Cache } from '@aws-amplify/cache';
 
 import { MESSAGE_TYPES } from '../src/Providers/constants';
 import * as constants from '../src/Providers/constants';
