+++ packages/pubsub/package.json
@@ -52,7 +52,6 @@
   ],
   "dependencies": {
     "@aws-amplify/auth": "5.5.3",
-    "@aws-amplify/cache": "5.1.3",
     "@aws-amplify/core": "5.5.2",
     "graphql": "15.8.0",
     "tslib": "^1.8.0",
