+++ packages/pubsub/webpack.config.js
@@ -6,7 +6,6 @@ module.exports = {
 		'graphql',
 		{
 			'@aws-amplify/auth': 'aws_amplify_auth',
-			'@aws-amplify/cache': 'aws_amplify_cache',
 			'@aws-amplify/core': 'aws_amplify_core'
 		}
 	],
